openrsync: implement the various --delete-* options for rsync parity
--delete-excluded may be paired with all of the other ones to indicate that
we should delete files that don't match exclude rules.

--delete-before, --delete-during, --delete-delay, --delete-after control
precisely when and where we collect the list of files to be deleted and
execute the removals. --delete-before is the pre-existing behavior, while
--delete-after is the after-transfer analog to it.  --delete-during and
--delete-delay both mean we simply check for deletions while we're updating
directories; in the former case, we delete entries on the spot and in the
latter we simply add them to a queue to be taken care of later.

X-From: ke/delopts