rules: implement MOD_NEGATE
The initial thought was that negate would be handled in rule_matched(), but
that's not quite the right place to do it; notably, the modifier is
documented to make the action take effect if the pattern does *not* match,
rather than negating the action (include -> exclude or vice versa).  This
means that we need to augment the pattern matching instead, and also that
we need to check further rules for matches if we don't match because of a
negation at play.