sender: stop reading fdin data after the second end-of-phase marker
We currently see a race in the test suite where we end up eating the end of
transmission marker because of how we defer the phase transition.  If there
are no updates needed, we may see progression through the second
end-of-phase and the final goodbye much faster than anticipated, leading to
a scenario where we end up eating the final goodbye marker and enqueueing it
before we have a chance to break out of the loop.

The solution is relatively straightforward: once we see two markers, we stop
reading regular data off of fdin.  We continue polling it and processing any
non-data messages that may come through in case they indicate we need to
terminate, but we need to avoid eating the final -1.