downloader: fix metadata not preserved on our temp file
If we're not preserving this metadata from the source, we really should try
to preserve it on the destination if the file did exist.  This would all
previously get clobbered by whatever permissions the temp file had, rather
than trying to keep it consistent.

Since the temp file now has the correct permissions, don't pass newfile to
rsync_set_metadata() unless the file actually didn't exist prior to the
update.