delete modes: check for any exempt children, skip
By necessity we now have to move directories to schedule deletion in the
post-order instead of pre-order so that we can apply any exclude rules
within.  Do it now for both the --delete-{delay,during} list construction
and for the --delete-{before,after} construction.

At some point we should investigate whether we can conslidate these back
together to remove redundancy.