download: offer an alternate path into cleanup
In a cleanup context, we really only care about cleaning up the partial
file -- we can skip all of the things that are not async-signal-safe, to
include unmapping/freeing memory and stdio.

While we're here, also document why we don't try to unlink the temp file if
we fail to rename it: partial files were requested, so they should decide
what to do with it.