rules: implement the cvs compat merge modifier
This uses a default name of .cvsignore so that :C as a rule can work, and it
also sets the n, w, and - merge modifiers.

Modifier parsing for this one had to be fixed; it would always hit the first
'C' entry and add just the MOD_CVSEXCLUDE bit.  The overlap is fine, the
'C' non-merge modifier doesn't make sense to be inherited, so we just set
both bits and unset the one that isn't applicable later on.