rules: implement processing side of dir-merge rules
Conceptually, it's relatively simple: each dir-merge rule may pick up 0 or
more directories with the named rule file.  Each of those is represented by
a struct merge_rule with a backpointer to the rule itself, its own array of
rules, and an indicator of whether it's inherited or not ('n' modifier).  We
then apply these in order when we hit a dir-merge rule as long as they're
inherited by the next directory level or the rules for the current
directory.

Next we'll implement the sender processing necessary.