rules: add infrastructure required for dir-merge handling
During flist traversal, we'll need to push/pop directories as we make our
way through them to do any of the needed maintenance.  We need to avoid
FTS chdir() while we're doing this so that we have stable access to the
dir-merge file.  Deeper subdirectories take priority over their parents, so
rules are inserted into their dir-merge chain in reverse order.