Avoid assuming implementation-defined behavior of malloc(0)
blkhash_set() may cause reallocarray with bset->blksz == 0, which is
implementation-defined and my either return NULL or a valid pointer that
cannot be dereferenced, but may be free()d.  Avoid relying on that, and
just return earlier if we don't have any blocks in the first place.

This quiets the transfer of empty files under such an implementation that
returns NULL, including at least some of LLVM's sanitizer interceptors.