rules: implement clear rules
The implementation is largely described within; since we're already scanning
rulesets when we enter and exit directories, we now keep track of the latest
clear rule we see in both directions and skip rule application entirely if
we had found one until we make our way back to it.

It's slightly convoluted, but we still have to start at the top because
rules that appear *after* the clear rule in parent dir-merge files still
need to be honored.