fargs: handle basic quoting in -e / --rsh
We don't need anything more than the bare basics; no escape sequences
supported, basic word-splitting.  If we come across a quotation mark, skip
to the end of it and continue word splitting in case we had a quote embedded
in the middle of an argument.  This is inline with how samba rsync seems to
handle the same, but it feels wrong- just note it in the -e description.