rules: fix --exclude/--include
We still try to parse a command, even with --exclude and --include rules.
This is wrong primarily because we're supposed to just accept the pattern
as-is, but also because we can end up with some nasty false positives.

This was surfaced by the test suite using an --exclude rule with a leading
space.  We'd call into parse_command() with len=0, which will cause
strncmp() to return 0 and we'll take the first command we see.  Fix that,
too, because a leading space in a normal command should result in an error.