downloader: lseek() past matching blocks
If we're resuming a transfer, for each block we'll receive a token
indicating which block should be written there.  If we're writing a block
that was already at that position (p->total is effectively our tracking of
the current file position), then there's no sense in dirtying up that
segment of the file.

The flush here is because, by default, we push writes into a 1MB write
buffer just to reduce write(2) calls a bit -- thus, we need to flush that
buffer to get our current position matching what we think it should be in
p->total, then we can safely seek ahead.  Odds are that, for most of the
scenarios where we can skip, there won't be anything buffered and we'll just
lseek() anyways.