Abstract away our --inplace checks a little bit
sess_is_inplace() can likely be called in more places, but for now it's
limited to just download_is_inplace().  The former may eventually be
expanded to cover, e.g., append mode, if we need --inplace to be implied
but don't want to set opts.inplace for whatever reason.

For the downloader, we need to be able to differentiate between "Is this an
in-place transfer?" and "Is this a *resumed* in-place transfer?"; for the
downloader, the answer may vary if we're, e.g., operating with a partial dir
and a partial file already existed in there.