receiver: implement the redo phase
If the final hash after the transfer doesn't match, then we actually need to
schedule it to be picked up in the second sending phase.  Right now we break
out at the end of the first and then the receiver closes out the second
phase, but with this patch we'll instead flag the uploader for re-entry,
re-enable polling, increase to the full long checksum size and start over
for those entries that needed redone.

The sender will actually already do the right thing, though it could perhaps
be a little more verbose about it.  There was one block of dead code that
prevented the redo phase from working against a client-sender smb rsync;
we'd send an extra ack that ended the redo phase prematurely.

This adds a new struct role that carries some role state with it.  For now
there's nothing actually unique to either sender or receiver role in there,
but a future change will refactor dlrename out of struct sess and into a
struct receiver that embeds struct role in it.