downloader: extend flist status to fully indicate current status
FLIST_COMPLETE indicates that the transfer is finished, but it hasn't been
finalized (e.g., moved into place or hardlinked).

FLIST_REDO indicates that we are going to redo the file.  If we're in the
redo phase, files will be in the REDO state until they've also FAILED or
until they end up complete- in the latter case, the FLIST_REDO bit is
stripped because we don't care that it was re-done; just that it's done now.

FLIST_SUCCESS is set once the file is completely done and in place, thus we
might send an acknowledgement to remove the file to the sender.  This may
happen immediately in the downloader, or it may be deferred until the end of
phase/transfer for --delay-updates.

FLIST_FAILED indicates that the file will not be completed, an unrecoverable
error has occurred.