io: add mechanism to tap out non-data messages
Other layers will register a handler to be notified when messages of a
specific type come in.  Currently we only support one handler at a time,
because I can't imagine needing more than one; it's left easy to extend if
we find a need later.

With this, we'll re-structure io_read_flush() a little bit.
mplex_read_remain gets zapped earlier on so that we don't need to worry
about it on every single return path until we get closer to the end, and we
do *some* things with zero-length payloads; we'll at least invoke a handler
if one exists for the type and let it decide if the short payload is an
error.  Receiving a zero-length payload with a logging/error tag is probably
an error on the remote side, but we should cope by proceeding onwards and
still aborting if it was an error tag.