receiver: implement the receiver-side of --remove-source-files
As we roll through each phase and at the end of each phase, we have to check
through each transferred file and see if we need to send a completion
notification so that the sender can remove it.  In the case of
delay-updates, we don't bother checking the flist until the end-of-phase
transition because files don't transition to the final state until the
update is actually processed.

As with the server-sender, the client-receiver needs to multiplex writes
since we need to be able to send SUCCESS messages back.