uploader: don't ignore errors from io_write_int_tagged
The error itself isn't all that important- if we've been disconnected or
something, for instance, we'll pickup the error in the next write which will
be something more critical.  In any event, we'll halt on error so that we
can retry later in case it was just a transient condition.