Fix rule_xfer
We can't use a quick heuristic to determine if it's an exclude rule because
we may have modifiers, so just drop that.  Pass the modifiers back to
rule_xfer() and use those in parse_rule_impl(), because otherwise the line
has been modified to just point directly at the pattern space.