Implement --sockopts
Set these on the socket we create to connect to the rsync:// daemon; in the
future, we'll also set them on the daemon socket when that's created.

smb rsync is a little vague about it, but the expected argument for
--sockopts is a comma-delimited list of name[=value] pairs, where name
corresponds to the literal name of socket option, including SO_ prefix, and
any amount of embedded whitespace may be found practically everywhere.

smb rsync does not require a value for some options (e.g., SO_{SND,RCV}BUF)
and it uses a default value of 1, but it's not clear to me that it makes any
sense to not require specificity there and I suspect in practice nobody is
relying on it.