downloader: some clean up for --delay-updates
In delayed_renames(), remove some redundancy by loading dlr->entries[i] once
into a local pointer.  As part of the cleanup mentioned next, only attempt
to rmdir(2) if curr->rmdir != NULL.

In rsync_downloader(), clean up some whitespace and create just a single
path through constructing the dir path that we'll be using to store our
staged updates in.  If we came up with the same directory as the previous
entry, just steal `rmdir` from it and set the previous entry's `rmdir` to
NULL to avoid an early deletion.  This is especially pertinent because
macOS will remove directories that only have an AppleDouble-formatted file
name in them, so it becomes more crucial that we don't try to rmdir(2) it
until the last file moves out.