Refactor our file spec normalization
At the end of fargs_parse(), we remove host names from all of our sources
before transmission.  Pull that part out into a separate function, because
we'll also need to do this normalization on the sink in order to support
writing to a daemon.

While we're here, rename one local variable (len -> hostlen) for some
clarity.