Don't setup filesfrom_fd in the client if we're the receiver
We sent the --files-from information over in the args in this case, so
there's nothing left to be done; the server will just read the file and send
us the flist information.

In the sender, only enable read multiplexing if we're the client to match
up with the 'else' block of this section.

(cherry picked from commit 7987ede683a318bdd201fbf9dc474ac23c646de9)