Fix --no-dirs and --no-relative without --files-from
We still need to actually reflect their usage in sess->opts, so still reset
opts.relative and opts.dir in the getopt loop.  This also lets us simplify
the --files-from path, as we just need to flip the options on if they're not
explicitly disabled.

While we're here, the smb rsync manpage also notes that -a doesn't imply
--recursive with --files-from, so note whether we have an implied recursive
or not and turn it off if we only had implied recursive.

(cherry picked from commit 9e42a796464b768157643a621045519daa0e26d1)