Don't bother actually taking the remote host for our filesfrom_host
We won't ever need it except for validating that it matches the one
specified in the source/sink, so just leave it as the empty string.  Others
will just test that it's not NULL to derive what they need.

(cherry picked from commit 46a97d34c1eecbf7eb3de71d9cffbf036d51ed01)