daemon: implement the "numeric ids" option
For this we have to turn numeric ids into a tri-state, because the daemon
may enable it without the client knowing and mess up the protocol -- we need
to send empty ID lists if we're doing numeric ids only based on the config,
and we still need to not remap anything if we're the receiving side.

Note that we still build the ident list if we're preserving uids/gids,
because we may need them for other purposes.  e.g., the unmerged --xattrs
that does POSIX ACLs would reference identities here.