socket: implement the client-side daemon auth protocol
If we get an AUTHREQD, then we need to request a password from the user and
hash it properly with the challenge issued, and write it back along with the
local username.  The username is taken from USER or LOGNAME, but should be
allowed in the rsync:// URI as well.

(cherry picked from commit 0f5d8b5cc02874559c5e1ffe6a06ec4f903b5b4d)