Switch --chmod to applying on the client-side
Previously we were applying it on the sending side, but it really needs to
be applied on whichever side of the transfer the client is on.  Pull the
application bits out into a flist_chmod() and apply it in both spots, and
move the parsing into rsync_client().

While we're here, fix <which>=<mode> -- we need to make sure we unset the
previous mask for the which bits that are set.