daemon: fix cfg_has_param
A module has a parameter even if it's specified in the global space, so we
should return true if that's the case.  The new logic is in-fact simpler, as
we don't need to skip deferring to global parameters in cfg_param_resolve().

This should fix specifying module parameters without a default at the global
level.  Parameters that just have a default set are not considered as being
'had'.

(cherry picked from commit 857691925cf4dfb37c751e7c884655dd8143155f)