daemon: implement symlink munging and non-munging behaivor
If munge symlinks is set, we need to prefix with /RSYNCD-MUNGED/ and if it's
not set but we're chrooted, we need to strip the symlink down to whatever
can fit inside of our module path.

For parity with smb rsync, refuse the client if the munging target exists
as a directory or a symlink to a directory.

(cherry picked from commit e8af37b16e346fb80cfea872a081a2c750bd6867)