daemon: warn if "dont compress" is set
Implementing it is more harmful than leaving it be -- it's not a security
feature, but just an optimization to avoid compressing files that are
already compressed.

If we implement it, then we're potentially disabling compression for the
entire transfer if the first file transferred happens to match the pattern.

If we don't implement it, at worst we're burning some more CPU cycles than
we should be, but with the trade-off that we're not necessarily burning a
lot more bandwidth than one might expect.

(cherry picked from commit e159d1926a89a8fc0cb43afc13a6b629fccab0da)