Prepare logging bits for daemon transfer logging
We need to do a round trip through an sbuf because we may be writing the
result to syslog instead of to a log file.  Build the output string up in
an sbuf, then use the normal log_write() facility to send it out.

main() already does an `rsync_set_logfile(stderr)` to direct output to
stderr, so remove `opts->logfile` without replacement.

While we're here, let print_7_or_8_bit() fail since it allocates its own
sbuf for formatting.  We can't reuse the one passed in even if it's not
NULL, since we need to expand a format string with `s` as the parameter.