downloader: fix send index remapping
This logic ended up duplicated by accident, so we could end up with an
erroneous remap as `idx` is the sendidx the first time, but the second time
it's actually the index into the flist.

Additionally, the bounds check at the beginning doesn't work if we pruned
entries near the end of the flist -- we can have a sendidx that's greater
than the new size of the flist, so we also need to scan the flist for that.