Fix sender-side of --copy-unsafe-links
In one case during flist generation, we were inadvertently falling through
and generating a second flist entry for the same entry; we'd end up with
one symlink and one directory, when we really should have only had the
directory.  This resulted in some intermittent errors where-in the receiver
seems to be requesting blocks for a symlink, but it's really just a mismatch
in flist view.

While we're here, fix up a couple spots where we're ignoring failures in
flist_gen_*() calls.