daemon: fix invocation from a socket initiator
We must grab a sockaddr for the socket that we were passed by launchd
since we'll need it for doing any host allow/deny processing, and we
must defer handling until after we read in our global options.
Reading the options twice here seems wasteful, because it kind of
is -- but it matches the smb rsync behavior.