openrsync: compression: Fix size of malloc'd buffer
Make sure the memory buffer includes the two extra bytes we need
to be able to deal with the 4 byte trailer that gets added during
the BLKSTAT_FLUSH phase of the state machine.

These 4 bytes might push user data to the next block.

TOKEN_MAX_DATA is 16383 bytes (reserves 2 bytes for a 14 bit flags field)

TOKEN_MAX_BUF is 2 bytes bigger to make sure we don't walk off the end
of our allocation when we assign cctx.next_out to cbuf+2.