openrsync: fix --delete with multiple trailing-/ paths
With multiple trailing '/' src, we'll be dumping multiple paths directly
into the destination directory rather than in subdirectories within the
destination directory's tree.  As a result, we end up erroring out in
flist_gen_dels() as we try to add the same directory multple times to the
hash table.

This change fixes numerous issues with this path, notably:
 - Fix --dirs mode by respecting FLIST_TOP_LEVEL coming from the other side;
    i.e., stop calling flist_topdirs() independently on the receive-side and
    set FLSTAT_TOP_DIR based on the received flist entry flags
 - Fix the logic in flist_topdirs() for the sender, --dirs should only set
    the TOP_LEVEL flag for entries ending with "/" or "/." while --recursive
    sets it for all directories.
 - Stop failing attempts to add duplicates to the hash table, we may
    legitimately encounter such a scenario if we have multiple srcs going
    directly in the destination directory.
 - Stop removing entries that we're needing to keep, the leading ./ must be
    dropped to line up with the flist path.

As an optimization, we'll drop duplicate top-level /. dirs in
flist_gen_dels() so that we don't walk the destination tree multiple times.