openrsync: fix remaining --delete bugs
In flist_gen_dels we had started stripping extraneous ./ at the begining
of paths before checking rules on them, but our flist entries may or may not
still have them -- strip it there, too.

Rules never apply to the root directory of a transfer, it simply would not
make sense.

While we're here, stop dropping stderr from test21 on the floor; this makes
it harder to figure out what's going on.