compression: fix issue when decompressing large blocks
When a large block size is used, MAX_CHUNK_BUF might not be enough
to hold all of the output. The input will have been comsumed, but
we need to process the rest of the output to leave of decompression
state in the correct place for the next chunk.

Since this is for a "fast forward", the blocks were "stored" and we
are copying from another file, we can just throw the buffer away, but
we need to drain all of the output first.

Don't break the loop until we have drained it all.