flist: sender: interpret the empty string as "."
smb rsync allows constructs like:

- rsync -a '' /tmp/holding: copy cwd to /tmp/holding
- rsync -a '' host:/tmp/holding: ditto
- rsync -a /tmp/holding '': copy /tmp/holding to cwd

It does *not* allow:
- rsync -a host:/tmp/holding '' (newer versions allow it with --old-args)

Capture all of this by reinterpreting it in the sender part of the flist,
which avoids accidentally working in the last case and fixes the second and
third case above.  The first case is fixed by the previous commit, which
ensures that we don't send '' to the remote host under any condition.

(cherry picked from commit 1325d512ba4a9b35b45321c76f8f6c390ab37de7)