openrsync: sync directory metadata properly
The pre-existing code for this is wrong, as the XXX to just use
rsync_set_metadata suggested.  The mtime logic was likely not right at all,
but more importantly it completely neglected to sync up directory
owner/group as needed.

Just use rsync_set_metadata_at() for this so that we get all of the metadata
we would normally get for every other type that we updated- it's just
deferred in case the update would have stripped our ability to add more
files or search the directory.

(cherry picked from commit 71cbd55a9928c21687e79d1907dd62abc2343026)