openrsync: unblock a potential hang in the sender
In the right circumstances, the sender can get blocked waiting on more block
metadata from the receiver.  The receiver is still waiting on the sender to
finish out parts of the current transfer, so it's blocked and unable to send
the metadata that the sender wants.

Relieve the strain by allowing block metadata to come in over multiple
reads, so that we can peel what we can off of the incoming socket but still
make progress on the write-side to avoid deadlocking.

(cherry picked from commit a83daa61920721c2488e19cca54099f53aa5f75f)